MB-34778: Make logger_map_mutex_ recursive.
The destruction of a BucketLogger calls
cb::logger::unregister_logger to remove the logger from the spdlog
registry. This caused a deadlock in spdlog when the background
flushing thread was the only owner of the shared_ptr. It would
attempt to destruct the BucketLogger but deadlock as it already held
the logger_map_mutex_ in spdlog::registry. Update spdlog registry to
make this mutex recursive which will allow the double locking by this
thread.

Change-Id: I5abed34e0f9614c55c0efef4126290b594aedc35