MB-34778: Defer destruction of shared_ptr in registry
If we are iterating on the loggers_ map in the registry and
the function is the last thing to own the shared_ptr then the
destruction (if it is a BucketLogger) will cause the element to
be erased from loggers_. This frees the iterator which causes ASAN
and TSan to report a heap use after free when we attempt to
increment the iterator. We can defer the destruction of the
shared_ptr to after we have advanced the iterator by storing the
latest one outside of the loop. When we overwrite it, (or when we
go out of scope) we will destroy the shared_ptr which is fine as we
no longer have an iterator pointing to it (erase does not invalidate
any other iterators).

Change-Id: If6ef0964d32ce8f03a9cbe34c84c0557f6d541bd